- Test suffixes migrates to plural version.
