+++ database/src/test/scala/za/co/absa/atum/database/runs/GetPartitioningCheckpointsIntegrationTests.scala
@@ -8,7 +8,7 @@ import java.time.OffsetDateTime
 import java.util.UUID
 import scala.util.Random
 
-class GetPartitioningCheckpointsIntegrationTest extends DBTestSuite{
+class GetPartitioningCheckpointsIntegrationTests extends DBTestSuite{
 
   private val fncGetPartitioningCheckpoints = "runs.get_partitioning_checkpoints"
