+++ model/src/test/scala/za/co/absa/atum/model/utils/SerializationUtilsUnitTests.scala
@@ -19,12 +19,12 @@ package za.co.absa.atum.model.utils
 import org.scalatest.flatspec.AnyFlatSpecLike
 import za.co.absa.atum.model.dto.MeasureResultDTO.{ResultValueType, TypedValue}
 import za.co.absa.atum.model.dto._
-import SerializationUtilsUnitTest.StringLinearization
+import za.co.absa.atum.model.utils.SerializationUtilsTest.StringLinearization
 
 import java.time.{ZoneId, ZoneOffset, ZonedDateTime}
 import java.util.UUID
 
-class SerializationUtilsUnitTest extends AnyFlatSpecLike {
+class SerializationUtilsUnitTests extends AnyFlatSpecLike {
 
   // AdditionalDataDTO
   "asJson" should "serialize AdditionalDataDTO into json string" in {
@@ -426,7 +426,7 @@ class SerializationUtilsUnitTest extends AnyFlatSpecLike {
 
 }
 
-object SerializationUtilsUnitTest {
+object SerializationUtilsTest {
   implicit class StringLinearization(val str: String) extends AnyVal {
     def linearize: String = {
       str.stripMargin.replace("\r", "").replace("\n", "")
