+++ server/src/test/scala/za/co/absa/atum/server/api/database/runs/functions/WriteCheckpointIntegrationTests.scala
@@ -17,7 +17,8 @@
 package za.co.absa.atum.server.api.database.runs.functions
 
 import org.junit.runner.RunWith
-import za.co.absa.atum.model.dto.{AdditionalDataSubmitDTO, PartitionDTO}
+import za.co.absa.atum.model.dto.MeasureResultDTO.{ResultValueType, TypedValue}
+import za.co.absa.atum.model.dto._
 import za.co.absa.atum.server.ConfigProviderTest
 import za.co.absa.atum.server.api.TestTransactorProvider
 import za.co.absa.atum.server.api.database.PostgresDatabaseProvider
@@ -27,32 +28,36 @@ import zio._
 import zio.test._
 import zio.test.junit.ZTestJUnitRunner
 
+import java.time.ZonedDateTime
+import java.util.UUID
+
 @RunWith(classOf[ZTestJUnitRunner])
-class CreateOrUpdateAdditionalDataIntegrationTest extends ConfigProviderTest {
+class WriteCheckpointIntegrationTests extends ConfigProviderTest {
 
   override def spec: Spec[TestEnvironment with Scope, Any] = {
 
-    suite("CreateOrUpdateAdditionalDataIntegrationSuite")(
-      test("Returns expected Right with Unit") {
-        val additionalDataSubmitDTO = AdditionalDataSubmitDTO(
+    suite("WriteCheckpointSuite")(
+      test("Returns expected Left with DataNotFoundException as related partitioning is not in the database") {
+        val checkpointDTO = CheckpointDTO(
+          id = UUID.randomUUID(),
+          name = "name",
+          author = "author",
           partitioning = Seq(PartitionDTO("key1", "val1"), PartitionDTO("key2", "val2")),
-          additionalData =  Map[String, Option[String]](
-            "ownership" -> Some("total"),
-            "role" -> Some("primary")
-          ),
-          author = "testAuthor"
+          processStartTime = ZonedDateTime.now(),
+          processEndTime = None,
+          measurements =
+            Set(MeasurementDTO(MeasureDTO("count", Seq("*")), MeasureResultDTO(TypedValue("1", ResultValueType.Long))))
         )
         for {
-          createOrUpdateAdditionalData <- ZIO.service[CreateOrUpdateAdditionalData]
-          result <- createOrUpdateAdditionalData(additionalDataSubmitDTO)
+          writeCheckpoint <- ZIO.service[WriteCheckpoint]
+          result <- writeCheckpoint(checkpointDTO)
         } yield assertTrue(result == Left(DataNotFoundException(FunctionStatus(41, "Partitioning not found"))))
       }
     ).provide(
-      CreateOrUpdateAdditionalData.layer,
+      WriteCheckpoint.layer,
       PostgresDatabaseProvider.layer,
       TestTransactorProvider.layerWithRollback
     )
-
   }
 
 }
