+++ server/src/test/scala/za/co/absa/atum/server/api/repository/PartitioningRepositoryIntegrationTests.scala
@@ -35,7 +35,7 @@ import zio.test._
 import zio.test.junit.ZTestJUnitRunner
 
 @RunWith(classOf[ZTestJUnitRunner])
-class PartitioningRepositoryIntegrationTest extends ZIOSpecDefault with TestData {
+class PartitioningRepositoryIntegrationTests extends ZIOSpecDefault with TestData {
 
   // Create Partitioning Mocks
   private val createPartitioningIfNotExistsMock = mock(classOf[CreatePartitioningIfNotExists])
@@ -78,7 +78,7 @@ class PartitioningRepositoryIntegrationTest extends ZIOSpecDefault with TestData
 
   override def spec: Spec[TestEnvironment with Scope, Any] = {
 
-    suite("PartitioningRepositoryIntegrationTest")(
+    suite("PartitioningRepositorySuite")(
       suite("CreatePartitioningIfNotExistsSuite")(
         test("Returns expected Right with Unit") {
           for {
